refactor consts.ts to consts.astro because of warning message when importing images
