+++ src/components/ContactUs.astro
@@ -6,7 +6,7 @@ import {
   ADDRESS_ZIP,
   PHONE_NUMBER,
   EMAIL_ADDRESS,
-} from "../consts";
+} from "../consts.astro";
 import ContactUsForm from "../components/ContactUsForm.astro";
 ---
