+++ src/layouts/BaseLayout.astro
@@ -1,5 +1,5 @@
 ---
-import { SITE_TITLE, SITE_DESCRIPTION } from "../consts.js";
+import { SITE_TITLE, SITE_DESCRIPTION } from "../consts.astro";
 import BaseHead from "../components/BaseHead.astro";
 import Header from "../components/Header.astro";
 import Footer from "../components/Footer.astro";
