+++ src/pages/about/our-team.astro
@@ -3,7 +3,7 @@ import DiscoveryFlightCTA from "../../components/DiscoveryFlightCTA.astro";
 import ContactUs from "../../components/ContactUs.astro";
 import BaseLayout from "../../layouts/BaseLayout.astro";
 import TeamMemberCard from "../../components/TeamMemberCard.astro";
-import { teamMembers } from "../../constants.astro";
+import { teamMembers } from "../../consts.astro";
 import Location from "../../components/Location.astro";
 ---
